Remove Errors namespace under Smithy::Client
