+++ gems/smithy-client/lib/smithy-client/service_error.rb
@@ -1,0 +1,57 @@
+# frozen_string_literal: true
+
+module Smithy
+  module Client
+    # The base class for all errors returned by a Smithy generated client.
+    # All ~400 level client errors and ~500 level server errors are raised
+    # as service errors. This indicates it was an error returned from the
+    # service and not one generated by the client.
+    class ServiceError < RuntimeError
+      # @param [Smithy::Client::HandlerContext] context
+      # @param [String] message
+      # @param [Structure] data
+      def initialize(context, message, data)
+        @code = self.class.code
+        @context = context
+        @message = extract_message(message, data)
+        @data = data
+        super(@message)
+      end
+
+      # @return [String, nil] The error code returned by the service.
+      attr_reader :code
+
+      # @return [Smithy::Client::HandlerContext] The context of the request
+      #  that triggered the remote service to return this error.
+      attr_reader :context
+
+      # @return [String] The error message returned by the service.
+      #  Defaults to the class name if no message is provided or can be parsed.
+      attr_reader :message
+
+      # @return [Structure] Additional data returned by the service.
+      attr_accessor :data
+
+      class << self
+        # @return [String, nil]
+        attr_accessor :code
+      end
+
+      # @return [Boolean] Returns `true` if the error is retryable.
+      def retryable?
+        false
+      end
+
+      # @return [Boolean] Returns `true` if the error is due to throttling.
+      def throttling?
+        false
+      end
+
+      private
+
+      def extract_message(message, data)
+        message || (data.message if data.respond_to?(:message)) || self.class.name.to_s
+      end
+    end
+  end
+end
