+++ gems/smithy-client/spec/smithy-client/plugins/retry_errors_spec.rb
@@ -100,7 +100,7 @@ module Smithy
           end
 
           let(:output) { Output.new(context: HandlerContext.new(config: config)) }
-          let(:service_error) { Errors::ServiceError.new(nil, nil, nil) }
+          let(:service_error) { ServiceError.new(nil, nil, nil) }
 
           let(:retry_strategy) { config.retry_strategy }
           let(:quota) { retry_strategy.instance_variable_get(:@quota) }
