+++ gems/smithy-client/spec/support/client_helper.rb
@@ -129,7 +129,8 @@ module ClientHelper
     def create_plan(module_name, model, options)
       plan_options = {
         module_name: module_name,
-        gem_version: options[:gem_version] || '0.1.0'
+        gem_version: options[:gem_version] || '0.1.0',
+        quiet: true
       }
       Smithy::Plan.new(model, :client, plan_options)
     end
