+++ gems/smithy/lib/smithy/templates/client/errors.erb
@@ -4,7 +4,7 @@
 
 module <%= module_name %>
   # When this service returns an error response, the SDK constructs and raises an error.
-  # These errors all extend <%= module_name %>::Errors::ServiceError < {Smithy::Client::Errors::ServiceError}
+  # These errors all extend <%= module_name %>::Errors::ServiceError < {Smithy::Client::ServiceError}
   #
   # You can rescue all errors using the ServiceError:
   #
@@ -16,7 +16,7 @@ module <%= module_name %>
   #
   # ## Request Context
   #
-  # ServiceError objects have a {Smithy::Client::Errors::ServiceError#context #context} method
+  # ServiceError objects have a {Smithy::Client::ServiceError#context #context} method
   # that returns information about the request that generated the error.
   # See {Smithy::Client::HandlerContext} for more information.
   #
@@ -29,13 +29,13 @@ module <%= module_name %>
   # Additionally, error classes are dynamically generated for service errors based on the error code
   # if they are not defined above.
   module Errors
-    extend Smithy::Client::Errors::DynamicErrors
+    extend Smithy::Client::DynamicErrors
 <% errors.each do |error| -%>
 
 <% error.docstrings.each do |docstring| -%>
     # <%= docstring %>
 <% end -%>
-    class <%= error.name %> < Smithy::Client::Errors::ServiceError
+    class <%= error.name %> < Smithy::Client::ServiceError
       # @param [Smithy::Client::HandlerContext] context
       # @param [String] message
       # @param [<%= module_name %>::Types::<%= error.name %>] data
