+++ projections/weather/lib/weather.rb
@@ -5,6 +5,9 @@
 require 'smithy-client'
 
 
+Object.const_set('Weather', Module.new) unless Object.const_defined?('Weather')
+
+# Provides weather forecasts.
 module Weather
   VERSION = '1.0.0'
 end
