+++ projections/weather/lib/weather/errors.rb
@@ -4,7 +4,7 @@
 
 module Weather
   # When this service returns an error response, the SDK constructs and raises an error.
-  # These errors all extend Weather::Errors::ServiceError < {Smithy::Client::Errors::ServiceError}
+  # These errors all extend Weather::Errors::ServiceError < {Smithy::Client::ServiceError}
   #
   # You can rescue all errors using the ServiceError:
   #
@@ -16,7 +16,7 @@ module Weather
   #
   # ## Request Context
   #
-  # ServiceError objects have a {Smithy::Client::Errors::ServiceError#context #context} method
+  # ServiceError objects have a {Smithy::Client::ServiceError#context #context} method
   # that returns information about the request that generated the error.
   # See {Smithy::Client::HandlerContext} for more information.
   #
@@ -27,10 +27,10 @@ module Weather
   # Additionally, error classes are dynamically generated for service errors based on the error code
   # if they are not defined above.
   module Errors
-    extend Smithy::Client::Errors::DynamicErrors
+    extend Smithy::Client::DynamicErrors
 
     # Error class for NoSuchResource.
-    class NoSuchResource < Smithy::Client::Errors::ServiceError
+    class NoSuchResource < Smithy::Client::ServiceError
       # @param [Smithy::Client::HandlerContext] context
       # @param [String] message
       # @param [Weather::Types::NoSuchResource] data
