+++ smithy-build.json
@@ -1,5 +1,10 @@
 {
     "version": "1.0",
+    "maven": {
+        "dependencies": [
+            "software.amazon.smithy:smithy-protocol-traits:1.56.0"
+        ]
+    },
     "projections": {
         "weather": {
             "plugins": {
